+++ run/src/main/java/com/guide/run/event/controller/EventSearchController.java
@@ -1,13 +1,10 @@
 package com.guide.run.event.controller;
 
-import com.guide.run.event.entity.dto.response.search.SearchAllEventsCount;
+import com.guide.run.admin.dto.SearchAllEventsCount;
 import com.guide.run.event.entity.dto.response.search.SearchAllEventList;
-import com.guide.run.event.entity.repository.EventRepository;
 import com.guide.run.event.service.EventSearchService;
-import com.guide.run.event.service.EventService;
 import com.guide.run.global.exception.user.resource.NotExistUserException;
 import com.guide.run.global.jwt.JwtProvider;
-
 import com.guide.run.user.repository.user.UserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
