+++ run/src/main/java/com/guide/run/event/entity/dto/response/search/SearchAllEventsCount.java
@@ -1,12 +1,12 @@
-package com.guide.run.event.entity.dto.response.search;
+package com.guide.run.admin.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
-@Getter
 @AllArgsConstructor
 @Builder
+@Getter
 public class SearchAllEventsCount {
     private int count;
 }
